Big logging improvements

+++ src/Clients/shared/src/services/usersServiceBase.ts
@@ -1,56 +1,59 @@
-import { inject } from "aurelia-framework";
-import { HttpClient } from "aurelia-fetch-client";
-import { EventAggregator } from "aurelia-event-aggregator";
+import { autoinject } from "aurelia-framework";
 
-import { HttpProxyBase } from "../../../shared/src/utils/httpProxyBase";
-
-import { AuthService } from "./authService";
+import { HttpProxy } from "../utils/httpProxy";
 import { LocalStorageBase } from "../utils/localStorageBase";
+import { ErrorLogger } from "./errorLogger";
 
-@inject(AuthService, HttpClient, EventAggregator, LocalStorageBase)
-export class UsersServiceBase extends HttpProxyBase {
+@autoinject
+export class UsersServiceBase {
   private readonly profileImageThumbRegex =
     /res.cloudinary.com\/personalassistant\/t_profile_thumbnail\/(development|production)\/users/g;
 
   constructor(
-    protected readonly authService: AuthService,
-    protected readonly httpClient: HttpClient,
-    protected readonly eventAggregator: EventAggregator,
+    protected readonly httpProxy: HttpProxy,
+    protected readonly logger: ErrorLogger,
     private readonly localStorageBase: LocalStorageBase
-  ) {
-    super(authService, httpClient, eventAggregator);
-  }
+  ) {}
 
   async getProfileImageUri(): Promise<string> {
-    const currentSrc = this.localStorageBase.getProfileImageUri();
+    try {
+      const currentSrc = this.localStorageBase.getProfileImageUri();
 
-    if (!navigator.onLine) {
-      return currentSrc;
-    }
+      if (!navigator.onLine) {
+        return currentSrc;
+      }
 
-    const src = await this.ajax<string>("users/profile-image-uri");
-    this.localStorageBase.setProfileImageUriLastLoad(new Date());
+      const src = await this.httpProxy.ajax<string>("api/users/profile-image-uri");
+      if (!src) {
+        throw new Error("Could not load profile image URI");
+      }
 
-    if (src !== currentSrc) {
-      this.localStorageBase.setProfileImageUri(src);
+      this.localStorageBase.setProfileImageUriLastLoad(new Date());
 
-      // Remove previous profile images from cache and add new one
-      caches.keys().then(async (cacheNames) => {
-        if (cacheNames.length > 0) {
-          const latest = cacheNames.sort().reverse()[0];
-          const cacheStorage = await caches.open(latest);
+      if (src !== currentSrc) {
+        this.localStorageBase.setProfileImageUri(src);
 
-          for (const request of await cacheStorage.keys()) {
-            if (request.url.match(this.profileImageThumbRegex)) {
-              cacheStorage.delete(request);
+        // Remove previous profile images from cache and add new one
+        caches.keys().then(async (cacheNames) => {
+          if (cacheNames.length > 0) {
+            const latest = cacheNames.sort().reverse()[0];
+            const cacheStorage = await caches.open(latest);
+
+            for (const request of await cacheStorage.keys()) {
+              if (request.url.match(this.profileImageThumbRegex)) {
+                cacheStorage.delete(request);
+              }
             }
+
+            await cacheStorage.add(new Request(src));
           }
+        });
+      }
 
-          await cacheStorage.add(new Request(src));
-        }
-      });
+      return src;
+    } catch (e) {
+      this.logger.logError(e);
+      throw e;
     }
-
-    return src;
   }
 }
